Add EventController and corresponding routing
A new EventController is added which includes a function that fetches all events and renders them using the Inertia package. A route to this controller is also added to the web routes file. The Vue.js version has been updated and a new Vue component for displaying events has been created.