The commit message does not match the provided diffs. The appropriate commit message should be:
Add categories data to event index

In the EventController, the Category model has been imported and all categories are fetched alongside all events when the invoke method is called. Also, in the Events/Index Vue component, a prop for categories has been added and a section to display each category has been implemented.